+++ src/main/java/com/borisfarber/instasearch/formats/Clazz.java
