+++ src/main/java/com/borisfarber/instasearch/formats/XmlDecompressor.java
