+++ src/main/java/com/borisfarber/instasearch/search/ZipSearch.java
@@ -148,7 +148,7 @@
 
          return Path.of(tempFile.toURI());
      }
-     
+
      @Override
      public void close() {
          executorService.shutdown();
